Has baseline snapshot
